Move the "last GPS ping" status into the header
Only for mini-schedule version. The stable version remains unchanged.
