+++ assets/css/properties_panel/_header.scss
@@ -4,11 +4,6 @@
   .m-route-variant-name {
     @include font-header;
   }
-
-  .m-close-button {
-    display: flex;
-    flex-direction: column;
-  }
 }
 
 .m-properties-panel__label {
@@ -80,7 +75,15 @@
   }
 }
 
-.m-properties-panel__header > button {
-  flex: 0 0 auto;
+.m-properties-panel__close-ping {
+  align-items: flex-end;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
   padding: 1rem;
+}
+
+.m-properties-panel__last-gps-ping {
+  @include font-small;
+  color: $color-font-grey;
 }
